nex-imgops/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nex-imgops"
version = "0.0.4"
description = "Image utilities / stream processor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "nexcli>=0.1.5",
    "numpy>=1.26",
    "Pillow>=10.3",
    "opencv-python>=4.9"
]

[project.entry-points.'nexcli.subcommands']
imgops = "nex_imgops:cli"
